Allocate crossings in batches on host as well. See below.
Little-to-no performance gain. Makes other things more complicated.
Didn't solve the initial issue, which is that some resolutions are
just too big to hold on the host (I thought it might work if the
bad_alloc error was due to a single allocated memory block being
too large, but alas...).